Integrate parser into NESSystem and load ROM into memory
I debated for a long time how best to transfer
the data retrieved by the ROM parser into system
memory. In the end, I decided that encapsulating the
ROM information into its own class and handing that
ROM object to the system was the simplest option.

As no protections are set for reaching the end of a ROM
(which should never happen in a game), out_of_range errors
are inevitable given that no CPU instructions have yet
been implemented. For our basic first demo, this should
be enough to show how the process should work, though
I would like to include registers and addresses along with
the reported opcodes.