Set Opcode flag whether PC may be changed
If the PC is changed by an instruction,
incrementing it at the end of the cycle will
throw the entire execution off. By delineating
operations that may manipulate the PC, we can
signal to the CPU whether or not to increment.

This means that it is the JMP/RET/BRN
instruction's responsibility to put the PC
in the correct location during the execute phase.