Execute moves program counter
Based on the documentation for the 6502, I'm
pretty sure the program counter is actually
changed by the individual instructions. Rather
than having some universal function that
is run for some instructions and not others
at the end of the loop, I think the simplest
recourse is to just let the execution stage
handle it. We have access to the opcode and
its AddressingMode anyway.

The CPU no longer progresses through the program.
Individual implementation of functions are
required to move forward.