Fix segfault in some compilers when allocating large ROM arrays
This error occurs depending on the size of the stack
allowed by a given compiler. This caused the program
to run successfully on my machine, but fail on
Michael's.

The PRG ROM and CHR ROM blocks can be very large
depending on the file. Originally, I had allocated
the 2d arrays on the stack when they really should
have been on the heap. The Rom object now contains
pointers to the PRG and CHR ROM data and deletes them
on deconstruction.