Implement stack operations
Stack pulling (popping) is strange with the 6502. It seems
that when the cpu pops from an empty stack, it "wraps"
around to 0x00. I have not implemented this behavior,
though it is something to keep in mind.

Reverted back to previous method of handling status flags
to more easily implement php() and plp() functions. This
can be changed to translating a temporary 8-bit value based
on the boolean flags, but for now, this method seems
simpler.