Implement shift/rotate instructions; get operand address, not operand itself
In order to implement in-place memory shifting or storing,
we need to save the operand address to be used later.
Rather than getting the operand in the fetch stage, we opt
to retrieve it (if needed) in the execute stage.

Also fixed memory storage instructions. Previously, the
program would attempt to retrieve an address to store by
accessing memory at the given address in the instruction.
In fact, THAT is the address that is supposed to be used.